Added error messages for login and moved loginand register component to components
