+++ src/app/components/login/login.component.html
@@ -1,0 +1,29 @@
+<body>
+<div class="main-div">
+<mat-card class="example-card">
+    <mat-toolbar-header>
+        <mat-card-title>Login</mat-card-title>
+    </mat-toolbar-header>
+    <mat-card-content [formGroup]="loginForm" class="flex-col">
+        <mat-form-field>
+            <input matInput placeholder="Email" type="text" formControlName="email">
+            <mat-error *ngIf="loginForm.controls.email.errors?.required">Email is required</mat-error>
+            <mat-error *ngIf="loginForm.controls.email.errors?.email">Email is invalid</mat-error>
+        </mat-form-field>
+        <mat-form-field>
+            <input matInput placeholder="Password" type="password" formControlName="password"/>
+            <mat-error>Password is required</mat-error>
+        </mat-form-field>
+    </mat-card-content>
+    <mat-card-action class="action-buttons">
+        <button mat-raised-button color="primary" [disabled]="loginForm.invalid" (click)="onSubmit()">
+            <!-- <span *ngIf="loading" class="spinner-border spinner-border-sm mr-1"></span> -->
+            Login
+        </button>
+        <button mat-raised-button color="primary" (click)="register()">
+            Register
+        </button> 
+    </mat-card-action>
+</mat-card>
+</div>
+</body>
