+++ src/app/components/login/login.component.scss
@@ -1,0 +1,21 @@
+.main-div {
+    height: 100vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+body {
+    background-color: azure
+}
+
+.flex-col {
+    display: flex;
+    flex-direction: column;
+}
+
+.action-buttons {
+    button {
+        margin: 10px;
+    }
+}
