+++ src/app/components/login/login.component.ts
@@ -1,9 +1,9 @@
+import { ActivatedRoute, Router } from '@angular/router';
 import { Component, OnInit } from '@angular/core';
-import { Router, ActivatedRoute } from '@angular/router';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { first } from 'rxjs/operators';
 
-import { AuthService } from '../services/auth.service';
+import { AuthService } from '../../services/auth.service';
+import { first } from 'rxjs/operators';
 
 @Component({
     templateUrl: 'login.component.html',
@@ -12,7 +12,6 @@ import { AuthService } from '../services/auth.service';
 export class LoginComponent implements OnInit {
     loginForm: FormGroup;
     loading = false;
-    submitted = false;
     returnUrl: string;
 
     constructor(
@@ -25,36 +24,33 @@ export class LoginComponent implements OnInit {
 
     ngOnInit() {
         this.loginForm = this.formBuilder.group({
-            username: ['', Validators.required],
+            email: ['', [Validators.required, Validators.email]],
             password: ['', Validators.required]
         });
 
         // get return url from route parameters or default to '/'
         this.returnUrl = this.route.snapshot.queryParams['returnUrl'] || '/';
+
+
+        // console.log(this.loginForm)
+        this.loginForm.valueChanges.subscribe(value => {
+            console.log(this.loginForm);
+        });
     }
 
-    // convenience getter for easy access to form fields
-    get f() { return this.loginForm.controls; }
-
     onSubmit() {
-        this.submitted = true;
-
-        // stop here if form is invalid
-        if (this.loginForm.invalid) {
-            return;
-        }
-
         this.loading = true;
-        this.authenticationService.login(this.f.username.value, this.f.password.value)
+
+        this.authenticationService.login(this.loginForm.value.email, this.loginForm.value.password)
             .pipe(first())
             .subscribe(
                 data => {
                     this.router.navigate(['/role']);
                 },
                 error => {
-                    //this.alertService.error(error);
                     this.loading = false;
-                });
+                }
+            );
     }
 
     register() {
