+++ src/app/components/register/register.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { first } from 'rxjs/operators';
-import { AuthService } from '../services/auth.service';
 
+import { AuthService } from '../../services/auth.service';
+import { Router } from '@angular/router';
+import { first } from 'rxjs/operators';
 
 @Component({
     templateUrl: 'register.component.html',
