--- src/app/login/login.component.html
@@ -1,44 +1,0 @@
-<body>
-<div class="main-div">
-<mat-card class="example-card">
-    <mat-toolbar-header>
-        <mat-card-title>Login</mat-card-title>
-    </mat-toolbar-header>
-    <mat-card-content>
-            <form [formGroup]="loginForm" (ngSubmit)="onSubmit()">
-                <table class="example-full-width" cellspacing="0">
-                    <tr>
-                        <td>
-                            <mat-form-field>
-                                <div class="form-group">
-                                        <input matInput placeholder="Username" type="text" formControlName="username" class="form-control" [ngClass]="{ 'is-invalid': submitted && f.username.errors }" />
-                                </div>
-                            </mat-form-field>
-                        </td>
-                    </tr>
-                    <tr>
-                        <td>
-                            <mat-form-field>
-                                <div class="form-group">
-                                    <input matInput placeholder="Password" type="password" formControlName="password" class="form-control" [ngClass]="{ 'is-invalid': submitted && f.password.errors }" />
-                                </div>
-                            </mat-form-field>
-                        </td>
-                    </tr>
-                    <div class="form-group">
-                        <button mat-raised-button color="primary" [disabled]="loading" class="btn btn-primary">
-                            <span *ngIf="loading" class="spinner-border spinner-border-sm mr-1"></span>
-                            Login
-                        </button>
-                    </div>
-                </table>
-            </form>
-    </mat-card-content>
-    <mat-card-action>
-            <button mat-raised-button color="primary" (click)="register()">Register</button> 
-
-    </mat-card-action>
-
-</mat-card>
-</div>
-</body>
