--- src/app/register/register.component.scss
@@ -1,10 +1,0 @@
-.main-div{
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-body{
-    background-color: azure
-}
